requirements.txt
streamlit
python-dotenv
tiktoken
chromadb
exa_py